Move test into its own file protected by built constraint
