+++ api/filesys/fsnode_test.go
@@ -11,7 +11,6 @@ import (
 	"sort"
 	"strings"
 	"testing"
-	"testing/iotest"
 )
 
 const content = `
@@ -205,40 +204,6 @@ type pathCase struct {
 	name   string
 	path   string
 	errStr string
-}
-
-func TestFileOps(t *testing.T) {
-	const path = "foo.txt"
-	content := strings.Repeat("longest content", 100)
-
-	fs := MakeFsInMemory()
-	f, err := fs.Create(path)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if _, err := fs.Open(path); err == nil {
-		t.Fatalf("expected already opened error, got nil")
-	}
-	if _, err := fmt.Fprint(f, content); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if err := f.Close(); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if err := f.Close(); err == nil {
-		t.Fatalf("expected already closed error, got nil")
-	}
-
-	f, err = fs.Open(path)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	defer f.Close()
-
-	if err := iotest.TestReader(f, []byte(content)); err != nil {
-		t.Fatalf("test reader failed: %v", err)
-	}
 }
 
 func TestAddDir(t *testing.T) {
